+++ app/src/main/kotlin/com/simplemobiletools/clock/activities/SettingsActivity.kt
@@ -1,5 +1,6 @@
 package com.simplemobiletools.clock.activities
 
+import android.content.Intent
 import android.os.Bundle
 import com.simplemobiletools.clock.R
 import com.simplemobiletools.clock.extensions.config
@@ -7,6 +8,7 @@ import com.simplemobiletools.clock.extensions.updateWidgets
 import com.simplemobiletools.clock.helpers.DEFAULT_MAX_ALARM_REMINDER_SECS
 import com.simplemobiletools.clock.helpers.DEFAULT_MAX_TIMER_REMINDER_SECS
 import com.simplemobiletools.commons.extensions.*
+import com.simplemobiletools.commons.helpers.IS_CUSTOMIZING_COLORS
 import com.simplemobiletools.commons.helpers.MINUTE_SECONDS
 import kotlinx.android.synthetic.main.activity_settings.*
 import java.util.*
@@ -20,6 +22,7 @@ class SettingsActivity : SimpleActivity() {
     override fun onResume() {
         super.onResume()
 
+        setupPurchaseThankYou()
         setupCustomizeColors()
         setupUseEnglish()
         setupAvoidWhatsNew()
@@ -32,7 +35,9 @@ class SettingsActivity : SimpleActivity() {
         setupSnoozeTime()
         setupVibrate()
         setupTimerMaxReminder()
+        setupIncreaseVolumeGradually()
         setupUseTextShadow()
+        setupCustomizeWidgetColors()
         updateTextColors(settings_holder)
         setupSectionColors()
     }
@@ -41,6 +46,13 @@ class SettingsActivity : SimpleActivity() {
         val adjustedPrimaryColor = getAdjustedPrimaryColor()
         arrayListOf(clock_tab_label, alarm_tab_label, stopwatch_tab_label, timer_tab_label, widgets_label).forEach {
             it.setTextColor(adjustedPrimaryColor)
+        }
+    }
+
+    private fun setupPurchaseThankYou() {
+        settings_purchase_thank_you_holder.beVisibleIf(config.appRunCount > 10 && !isThankYouInstalled())
+        settings_purchase_thank_you_holder.setOnClickListener {
+            launchPurchaseThankYouIntent()
         }
     }
 
@@ -56,7 +68,7 @@ class SettingsActivity : SimpleActivity() {
         settings_use_english_holder.setOnClickListener {
             settings_use_english.toggle()
             config.useEnglish = settings_use_english.isChecked
-            useEnglishToggled()
+            System.exit(0)
         }
     }
 
@@ -149,6 +161,14 @@ class SettingsActivity : SimpleActivity() {
         }
     }
 
+    private fun setupIncreaseVolumeGradually() {
+        settings_increase_volume_gradually.isChecked = config.increaseVolumeGradually
+        settings_increase_volume_gradually_holder.setOnClickListener {
+            settings_increase_volume_gradually.toggle()
+            config.increaseVolumeGradually = settings_increase_volume_gradually.isChecked
+        }
+    }
+
     private fun setupUseTextShadow() {
         settings_use_text_shadow.isChecked = config.useTextShadow
         settings_use_text_shadow_holder.setOnClickListener {
@@ -168,5 +188,14 @@ class SettingsActivity : SimpleActivity() {
 
     private fun updateTimerMaxReminderText() {
         settings_timer_max_reminder.text = formatSecondsToTimeString(config.timerMaxReminderSecs)
+    }
+
+    private fun setupCustomizeWidgetColors() {
+        settings_customize_widget_colors_holder.setOnClickListener {
+            Intent(this, WidgetDateTimeConfigureActivity::class.java).apply {
+                putExtra(IS_CUSTOMIZING_COLORS, true)
+                startActivity(this)
+            }
+        }
     }
 }