+++ app/src/main/kotlin/com/simplemobiletools/clock/adapters/ViewPagerAdapter.kt
@@ -10,6 +10,7 @@ import com.simplemobiletools.clock.fragments.StopwatchFragment
 import com.simplemobiletools.clock.fragments.TimerFragment
 import com.simplemobiletools.clock.helpers.TABS_COUNT
 import com.simplemobiletools.clock.helpers.TAB_ALARM
+import com.simplemobiletools.clock.helpers.TAB_CLOCK
 import com.simplemobiletools.clock.helpers.TAB_TIMER
 import com.simplemobiletools.commons.models.AlarmSound
 
@@ -37,11 +38,15 @@ class ViewPagerAdapter(fm: FragmentManager) : FragmentStatePagerAdapter(fm) {
         else -> throw RuntimeException("Trying to fetch unknown fragment id $position")
     }
 
+    fun updateClockTabAlarm() {
+        (fragments[TAB_CLOCK] as? ClockFragment)?.updateAlarm()
+    }
+
     fun updateAlarmTabAlarmSound(alarmSound: AlarmSound) {
-        (fragments[TAB_ALARM] as AlarmFragment).updateAlarmSound(alarmSound)
+        (fragments[TAB_ALARM] as? AlarmFragment)?.updateAlarmSound(alarmSound)
     }
 
     fun updateTimerTabAlarmSound(alarmSound: AlarmSound) {
-        (fragments[TAB_TIMER] as TimerFragment).updateAlarmSound(alarmSound)
+        (fragments[TAB_TIMER] as? TimerFragment)?.updateAlarmSound(alarmSound)
     }
 }