+++ CHANGELOG.md
@@ -1,6 +1,47 @@
 Changelog
 ==========
 
+Version 4.3.0 *(2018-09-19)*
+----------------------------
+
+ * Rewrote the widget to scale text size dynamically on Android 8+
+ * Couple stability and translation improvements
+
+Version 4.2.0 *(2018-08-10)*
+----------------------------
+
+ * Show a timer notification if the app gets paused while it runs
+ * Couple other UX improvements
+
+Version 4.1.0 *(2018-07-08)*
+----------------------------
+
+ * Adding a toggle for gradual volume increasing
+ * Rework the Alarm screen to use swipe to left/right for snooze/dismiss
+ * Couple other UX improvements
+
+Version 4.0.1 *(2018-05-27)*
+----------------------------
+
+ * Improve alarm reliability
+ * Play the timer sound at the Alarm sound stream
+ * Couple stability and UX improvements
+
+Version 4.0.0 *(2018-05-16)*
+----------------------------
+
+ * Allow changing the app launcher icon color
+ * Use bigger text size at widget, when made higher
+ * Show the next alarm at the clock fragment, if set
+ * Add a settings button for customizing widget color without recreating it
+
+Version 3.1.3 *(2018-04-11)*
+----------------------------
+
+ * Do not reset stopwatch and timer at device rotation
+ * Fix notification sounds on Android 8+
+ * Adding a No Sound option at notification sounds
+
 Version 3.1.2 *(2018-03-29)*
 ----------------------------